pyproject.toml
[project]
name = "mcp-docs-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
]
